+++ Complexity/Application/DataBaseExtensions.cs
@@ -22,12 +22,11 @@ namespace Application
                     .ToDictionary(
                         topic => topic.Id,
                         topic => new TopicProgressEntity(
-                            topic.Id,
                             topic.Levels
                                 .Take(1)
                                 .ToDictionary(
                                     level => level.Id,
-                                    level => level.ToProgressEntity()))));
+                                    level => level.ToProgressEntity()), topic.Id)));
 
             return userRepository.FindById(userId) ?? userRepository.Insert(new UserEntity(userId, progress));
         }
